requirements.txt
cv2
numpy
pillow
django=1.10
